+++ packages/ispect/lib/src/features/log_viewer/presentation/widgets/app_bar/app_bar.dart
@@ -224,10 +224,6 @@ class _ISpectAppBarState extends State<ISpectAppBar> {
   void _showFilterSheet(BuildContext context) {
     showISpectSheet<void>(
       context,
-      fitContent: false,
-      initialChildSize: 0.5,
-      minChildSize: 0.3,
-      maxChildSize: 0.8,
       topOnlyRadius: true,
       routeSettings: const RouteSettings(name: 'ISpect Filter Sheet'),
       builder: (context, scrollController) => ISpectFilterSheet(